requirements.txt
pytest
pygame
torch
opencv-python
numpy
click
pillow